requirements.txt
Flask==3.0.0
GitPython==3.1.40
gunicorn==21.2.0
APScheduler==3.10.4

